Fix more bugs that blocked successful compilation
The biggest one was the one that emptied any file that find_and_replace was run on --- this was due to the fact that the file was replaced by open(...,"r").read() after open(...,"w") already opened the file and cleared it. This means that the string that is written back will be the empty string.