Move the find_and_replace in QEMU out of the `init` function, and to the `configure` function
Currently, the class will try to run the find and replace step(s) when initializing. However, if the QEMU source code was not downloaded ahead of time, running `qemu=download` will fail --- the files don't exist to begin with.

However, if you are running configure, we are assuming the files exist at this point, so it is safe to run.